+++ changelog.md
@@ -1,5 +1,17 @@
 # Enable bbPress for Block Themes changelog
 
+## 1.0.1, 20250804
+
+### Fixed
+- Forum search and tags issue (https://wordpress.org/support/topic/problem-with-bbpress-forum-search/)
+
+### File updates
+	changelog.md
+	enable-bbp-block-theme.php
+	readme.txt
+	includes/Compatibility.php
+
+
 ## 1.0.0, 20250213
 
 - Initial release.